pyproject.toml
[build-system]
requires = ["poetry==0.12.17"]
build-backend = "poetry.masonry.api"


[tool.poetry]
name = "yolo-fc"
description = "saturation forecaster"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "~3.7.3"
django = "^3.0"
django-split-settings = "^1.0.0"
django-axes = "^5.0"
django-health-check = "^3.10"
django-feature-policy = "^3.0"
psycopg2 = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.1"
dump-env = "^1.0"
bcrypt = "^3.1"
django-extensions = "^2.2"
python-dateutil = "^2.8"
django-bulma = "^0.6.0"
django-braces = "^1.13"
jupyter = "^1.0"
django-pandas = "^0.6.1"
#tensorflow = "^2.1.0"
scikit-learn = "^0.22"
tqdm = "^4.35"
pyoo = "^1.4"
openpyxl = "^2.6"
model_bakery = "^1.0.2"
django-enumfield = "^2.0.1"
django-import-export = "^2.0"
Babel = "^2.8.0"
django-sns-view = "^0.1.2"
boto3 = "^1.11.1"
tablib = {extras = ["xlsx"], version = "^1.0.0"}
graphene-django = "^2.8.0"
django-filter = "^2.2.0"
funcy = "^1.14"
xlrd = "^1.2.0"
django-graphql-jwt = "^0.3.0"
six = "^1.14.0"
django-graphql-social-auth = {extras = ["jwt"], version = "^0.1.4"}
itsdangerous = "^1.1.0"
stripe = "^2.43.0"
paypalrestsdk = "^1.13.1"
paypal-checkout-serversdk = "^1.0.1"
pytest-bdd = "^3.2.1"
dynamic-conf = "^0.5.0"
django-cors-headers = "^3.2.1"
django-admin-list-filter-dropdown = "^1.0.3"
django-hint = "^0.1.5"
huey = "^2.2.0"
redis = "^3.4.1"
sentry-sdk = "^0.14.3"
shortuuid = "^1.0.1"

[tool.poetry.dev-dependencies]
black = {version = "*", allow-prereleases = true}
django-debug-toolbar = "^2.0"
django-querycount = "^0.7"
nplusone = "^1.0"
flake8-pytest = "^1.3"
flake8-django = "^0.0.4"
pytest = "=5.3.5"
pytest-django = "^3.5"
pytest-cov = "^2.7"
pytest-randomly = "^3.0"
pytest-deadfixtures = "^2.1"
pytest-testmon = "^1.0.2"
pytest-timeout = "^1.3"
pytest-env = "^0.6.2"
pre-commit = "^2.2.0"
xenon = "*"
mypy = "^0.770"
ipython = "*"
sphinx = "^2.1"
sphinx-autodoc-typehints = "^1.7"
doc8 = "^0.8"
yamllint = "^1.15"
safety = "^1.8"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
invoke = "^1.2"
pytest-xdist = "^1.29"
pylint-django = "^2.0.13"
autoflake = "^1.3.1"
better_exceptions = "^0.2.2"
gql-query-builder = "^0.1.5"
pdbpp = "^0.10.2"
sniffer = "^0.4.1"
MonkeyType = "^19.11.2"
pyinotify = {version="^0.9.6", platform="linux"}
django-stubs = "^1.5.0"
wemake-python-styleguide = "^0.14.0"
pytest-mock = "^3.0.0"
model_bakery = "^1.1.0"
mixer = {git = "https://github.com/lucomsky/mixer", branch="faker-version-update"}

[tool.black]
target-version = ["py37", ]
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | builds
    | dist
    | migrations
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
